third_party/blink/web_tests/fast/history/form-submit-in-frame-via-onclick-expected.txt

Tests that form submissions (via an onclick handler) inside an iframe generate history entries.

--------
Frame: 'the-frame'
--------
This page is the target of a redirect.

PASS: History item count should be 2 and is.


============== Back Forward List ==============
        (file test):fast/history/form-submit-in-frame-via-onclick.html
            (file test):fast/history/resources/form-with-onclick-submit.html (in frame "the-frame")
curr->  (file test):fast/history/form-submit-in-frame-via-onclick.html
            (file test):fast/history/resources/redirect-target.html#2 (in frame "the-frame")
===============================================
